Check whether we have manually set a dataset before setting the default one when constructing a views query.
It's possible to specify a dataset via a view filter, so we shouldn't force the user over to our default dataset in that condition. We needed to add this code as we previously added code to force some value being set for the site (to prevent always getting Eldis results unless we manually set a site filter on the view).